Add PTE_A flag to page table entries in riscv.h and vm.c
This commit adds the PTE_A (access flag) to the page table entries in the riscv.h and vm.c files. The PTE_A flag is used to track whether a page has been accessed or not. This flag is set when a page is accessed and can be cleared by the sys_pgaccess system call. The addition of this flag enhances the functionality of the page table and allows for more efficient memory management.